backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
pydantic
sqlalchemy
aiosqlite
python-dotenv
weasyprint
reportlab
jinja2
python-multipart
